Improving efficiency of WildcardPattern, adding getMatches(..) method
The asterisk is the most expensive part of a WildcardPattern matching
scan: it initiates a new loop to explore 0-n many possibilities (where n
is the length of the unmatched chars in the input string). And if there
are multiple asterisks: each one may compound the complexity.

This change reduces the complexity if there is a single asterisk to
zero, and will improve the efficiency of patterns with multiple
asterisks.