Continuation of previous commit
Somehow the window was being created, but the content wasn't being
applied correctly. I tried a few things, but this is the trick that
appeared to fully resolve the problem.

(The steps are the same as the previous commit: tab focus to password
field in popover demo, toggle caps lock repeatedly.)