Replacing ObservableList
The previous model wasn't bad. In fact: it could support at least one
scenario this model couldn't. (Specifically: once listeners were being
notified of a change, the previous implementation allowed other threads
to read the list.) But the previous implementation had some design
flaws, and its semaphore-based model could still result in deadlock.

This implementation is simpler, and it doesn't have the same lofty
ambition as the previous implementation. Listeners can modify the
ObservableList, and that's OK.

Also this introduces a more specific set of events to describe changes.