Removing TODO to merge LIKE operators
I spent a couple of days on this, but I think this is too complex.

Consider the patterns:
PRE*
*FIX

You should be able to AND them together as PRE*FIX. Any String that
matches both the original patterns should also match PRE*FIX

But this basic idea did not translate to "A*" and "*A". These should
combine as "A*A". But this means the String "A" matches the two original
patterns but NOT the new pattern: so this merge failed.

That's the last realization I had before I decided to pass on this
optimization.