Adding alternative .equals(..) comparison
This will be important to confirm split/joins are really joining
operators the way we expect (in a later stage).

This commit required rethinking how & when we simplify things.
Basically: now we don't attempt to simplify anything as it is first
constructed. We only do that when converting to a canonical form.