New equals() implementation based on testing values
By uncommenting out a TODO in the unit tests I exposed a flaw in the
boolean logic used to simplify expressions.

At the end of the day the following expression stumped the code:

(a && b && c && d) || 
(j) || 
(!a && !j) || 
(!b && !j) || 
(!c && !j) || 
(!d && !j)

This should always evaluate to 'true', because the first two products
((a && b && c && d) || j) when negated equal the remaining products, so
this is basically "x || !x"

But I couldn't figure out an efficient way to implement the factoring
required that conclusion. (I did apply some complex iteration of
subsets, but that made the unit tests jump to 20+ seconds.) And more
importantly: it drove home the conclusion that this logic is all too
precarious.

Instead with this commit:
We switch to a new brute-force model for testing equality. This
basically compares the truth table for two operations to confirm that
they'll produce the same results given the same input. Since an Operator
usually has a pretty small number of inputs this calculation is not very
intense. (It will scale badly, though, if we ever start to have
Operators with several dozen inputs.)

This had the added benefit of pointing out where a couple of unit tests
were wrong. So this new implementation caught problems the old one
didn't catch, which is a good sign.

I kept the "canonical" methods (renamed as "sumOfProducts") primarily to
help with the split/join feature, but I threw out the boolean algebra
used to simplify expressions.