Make Operator.equals(Object) use a strict comparator
In writing up the documentation: I tried to explain the subtle nuances
of hashCode() and equals(..), and I decided instead just to make the
default .equals implementation use a strict comparison.

This required retooling a lot of unit tests, and while I was there I
switched to String parsing to set up complex expressions.

A few things don't work like they used to, so this introduced a few TODO
notes in the unit tests. This is a good backup point, though.

(Also this adds several paragraphs of javadoc across a few classes.)