Issue #43: checkbox and labels in lead column weren't right-aligned
This appears to resolve the issue on Mac. I'll test Windows next.

Also this refactor appears to resolve an unreported issue: the baseline
for certain rows that contained wrapped radiobuttons could be 1 pixel
off.

I removed the notion of "horizontal weight" and instead just off a
stretch-to-fill boolean. (I've never really wanted a weight other than
zero or one so far... so why make it more complex than a boolean?)