Refactor QButtonUIs
This implementation extends BasicButtonUI and manages the custom "look"
as a javax.swing.Border. This lets us inherit the default layout/sizing
mechanisms in BasicButtonUI.

This effectively gets us:
1. HTML support
2. Baseline support
3. Correct vertical placement
(And maybe other things the BasicButtonUI handles that I'm not aware
of?)

What we lost is:
1. The ability to support custom arbitrary button shapes. (But I added
support for circles, which is the only custom shape I ever really use.)
2. Some flexibility calculating the preferred size. Now the Borders are
a fixed size. Previously we dynamically calculated the "border" (without
officially calling it that), and for really rounded UIs like the
RoundRectButtonUI: we used a variable width based on the height.

Also this included refactoring lots of supporting classes (like the
AnimationManager). And I pulled several QButtonUI subclasses and
UIEffects. If a solid use case ever presents itself I don't mind
bringing them back, but those were from a decade ago before UIs
emphasized simple/flatter looks.