Issue #46: animating collapsing panels was too slow
Previously we just relied on a timer and incremented every time the
timer had the chance to run. On slower machines this resulted in the
animation taking a variable amount of time.

Now we have a target animation time, so even if we skip a few frames:
the animation will end at a fixed time.