The AngleSliderUI might not render initially
I tried to figure out why the ComponentListener wasn't being notified
when the bounds were changed, but I couldn't figure out exactly what was
happening. (We push a ComponentEvent to the system event queue... and
then... ?)

So instead I just rewrote the AngleSliderUI class to automatically
recalculate geometry when the dimensions change.