Optimizing loops
This assumes the majority of pixels in a shadow images are identical to
the preceding pixel. (That is: transparent follows transparent, black
follows black.) So as we start to process a new kernel: compare the
newest pixel with our previous calculated value. If they match:
immediately iterate forward and look for more interesting pixel data.

Now rendering performance resembles:

Kernel	FastShadowRenderer	GaussianShadowRenderer
1.0 	35.0 	64.0
2.0 	36.0 	66.0
3.0 	22.0 	69.0
4.0 	23.0 	74.0
5.0 	23.0 	119.0
6.0 	24.0 	87.0
7.0 	25.0 	91.0
8.0 	28.0 	101.0
9.0 	29.0 	105.0
10.0 	31.0 	153.0
11.0 	27.0 	119.0
12.0 	28.0 	129.0
13.0 	28.0 	132.0
14.0 	28.0 	179.0
15.0 	30.0 	148.0