Fixing errors related to fractional kernels in shadows
While developing shadows for text boxes it came to my attention that
fractional kernels could throw an array index RTE for an opaque image.

The problem was that the vertical rendering pass could create values >
255. When the horizontal pass crunched those numbers, it ended up
consulting the LUT for a value that was outside of the table.

The key elements the unit tests were missing were:
1. Tests that included fractional kernels.
2. Tests that include opaque pixels.

This commit may go overboard in adding new tests to address this, but I
really want to avoid any more nasty surprises going forward. (It's one
thing if a bug comes up that shows that shadows are just a little
numerically off. It's much more severe if it comes up that rendering
shadows is throwing RuntimeExceptions.)