Adding clipping (overflow:hidden) for blocks (divs)
This is the first change that affects the layout of pages. Until now
every enhancement tried to keep elements positioned exactly the same as
Swing's default engine.

The unit tests also uncovered some double-painting, so how we paint the
backgrounds was tweaked a little bit. Double-painting opaque colors is
harmless, but if we did that to translucent colors we'd get the wrong
output.

Also in Graphics2D: we need to return an empty shape if the clipping is
empty. When we return null some Swing View code throws a NPE, because it
assumes the clipping is never null. (But returning an empty shape is
OK.)