Adding a new unit test re background-color
We have to make sure h1 tags use the background-color, but at the same
time we have to make sure the div tag (specifically
testOverflow_hidden_div) only applies the background-color once. (In
some unit test failures: the background was being painted twice. Which
may be harmless for opaque colors, but is not harmless for translucent
colors.) This commit appears to satisfy both unit tests.

I'm still fuzzy on exactly why/how background-color gets special
treatment in the StyleSheet.