Fixing bug with tiling text
The symptoms and problem are hard to describe. Symptoms:
Open the QHTML demo using the default window size.
The demos are all so large that they should wrap "LOREM IPSUM" onto two
separate lines, but instead we see two consecutive lines that contain
both words.

(The behavior if you try to make selection is even weirder.)

The problem was primarily with the QInlineView. If we had a QInlineView
X, it created an instance of the QViewHelper Z that referenced X.
Then somewhere (outside of our code) X was cloned. The clone (J)
included a reference to QViewHelper Z, which included a reference to our
original X.

So when J attempted to paint itself, it ended up calling X's
paintLegacyCss2 method. J was updated to include the appropriate
wrapping information, but X was not. So rendering was just a mess.

This could have been resolved by fixing the clone() method, but instead
I made QViewHelper static/transient. (So you access it with one public
static method. It happens to instantiate in that method in a throw-away
object, but that doesn't matter to the caller.)