Refactor how we serialize some awt beans
The previous code was designed for serialization. This change abstracts
that basic idea so we convert everything to a map. This map can then be
used for serialization AND equals() AND toString() AND hashcodes.

All this originally started with a TODO note in QViewRenderer about
Operation#equals(Object) not being sufficient to compare similar sets of
Operations.

Previously Operation#equals would serialize objects A and B into byte
arrays to check for equality. This was wasteful (and very hard to
inspect in the debugger). This new implementation just lets us compare
the Maps the converters produce.

There's still some room for potential efficiency gains: we have to write
two *entire* maps to begin comparing them. We could, in theory, use some
streaming or iterative approach so we could compare each attribute as it
becomes available and abort at the first noticeable difference. (This
might be esp useful for complex data, like large images.)