Let most recent CSS properties "win"
When two CSS properties conflict, the most recently defined one should
take priority. I think the BorderRenderingConfiguration is the only case
in this project where this principle actually matters, because that's
the only place (so far?) where we have constituent properties. (That is:
the "border" property and "border-color" property have the potential to
contain conflicting instructions.)

I tried other (simpler) approaches (like modifying the QStyleSheet's
Style to incorporate this project's CssValue's), but that kept falling
apart for some reason. At least I understand this approach, even if it's
clunky.q