Adding JPEGMetaDataListener
The previous implementation let you pass in a boolean to the parser: do
you want to fetch the thumbnails or not? Yes or no.

This new implementation lets a JPEGMetaDataListener decide on the fly
whether it wants a specific thumbnail. (For ex: you may only be
interested in a thumbnail if it's at least 100px wide.)