Rewriting JPEGMetaDataDemo, minor code cleanup
Making the listener implement AutoCloseable (or even Closeable) was a
weird/bad choice. The listener doesn't "close" like an InputStream or
other resource would; it can immediately fire up again for a new file
(depending on how you're using the reader). So I use "startFile" and
"endFile" instead just so the listener is notified. But it shouldn't,
for example, support a try-with-resources block.