Addressing an IllegalArgumentException with adjacent positions
When this happens we need to subtly differentiate adjacent positions.

(Now if there are THREE adjacent positions... that could still throw an
exception, and maybe that's best? If we find an example like that then
maybe we can judge then and there what the best response is / what the
intention of that CSS is...)