Encourage the user to make mistakes
Previously I half-heartedly tried to censor certain controls that were
going to fail. (For ex: you can't change a window's type once it's
displayed.)

It would be cumbersome to be sure I caught all such rules, though. Plus
part of the point of this demo is to let the user try for themselves
what works and what doesn't. (If anything changed in the JVM over time:
I wouldn't want my window demo to be stuck with some outdated rules I
hardwired into the UI.)

Instead: let's offer as few restrictions as possible, and roll back changes when we detect a problem. If this UI were for regular users I'd want to minimize the chance of encountering errors, but the target user here is a developer who will understand most errors.