Refactoring pixel packages
Now ImageLoader and Scaling let you request an image type, including
TYPE_DEFAULT to indicate you want whatever is fastest/easiest.

This refactor is not complete (I left some TODO notes in the code), but
it's spiraled into a lot more work than I intended and this is a good
backup point.