I think (?) this improves performance
I saw some bad EDT lag after I walked through most of the demos (including the HTML) in a session that included use of the MagnificationPanel. A stacktrace in the console suggested its listeners were being too needy somehow. I didn't explore this thoroughly, but plugging in the DescendantListener should achieve the same basic goal and hopefully it's a little more responsible.

If this persists I may have to either fix the DescendantListener or rethink this class.