Replacing ShapeBounds#getBounds
The new implementation is copied and pasted from new code introduced for
open-jdk ticket 8176501.

This also adds a unit test that failed in the previous implementation.

The old method getEdgePoints(..) will still fail for degenerate cubics.
I moved that so it's now a private method in BorderRendering -- which is
the only other class that used it. I have not thoroughly investigated
BorderRendering to make sure it's usage of getEdgePoints() is safe, but
my guess is its usage is actually much more restricted so it may be OK.