Adding PixelSourceImageProducer
I experimented with trying to make a java.awt.Image that relied on a PixelIterator.Source, but it turns out if you try to render a custom Image subclass the default Graphics2D classes just give up (they return false).

I pinged the client dev list about this, but even if they agree this is a problem (and even if it gets resolved) that solution would require, what, JDK22? I'm trying to keep this codebase compatible with older JDKs for now.