Abandoned ImagePixelIterators should auto-close their helper threads
Until recent commits: we used a thread pool with a fixed size. It was easy to max out that thread pool, and so a few orphaned production threads could really mess up ALL future production runnables. (They would eventually get their chance to run when production timed out, but could conspicuously take a long time.) All that is to say: aborting threads *used* to be a bigger deal.

Now we don't use a thread pool, and every image producer gets its own thread. But still: it's good practice to try to keep our thread count down. If we can tell an iterator is abandoned: let's tell the helper thread to cancel ASAP.