Issue #106: Adding rough draft of GestureListenerDemo
I have to set this aside for now, but this is an OK start. I think (?) it'll be impossible to use this demo without modifying the runtime permissions. So:
A. Triple-check this, maybe ask around. See if there's anything I'm not thinking of/aware of.
B. Make the demo savvy enough to warn users when we don't have access to the appropriate listeners.