Moving some unit tests to image module
I also deleted JPEGThumbnailImageReader.

The unit test showed this could be very problematic:

Depending on the order of execution, when JPEGMetaData eventually called ImageIO to read the inner JPEG image: it might settle on the pump class JPEGThumbnailImageReader. This class is incapable of reading a whole JPEG, and it throws an UnsupportedOperationException if you attempt it.

This made me nervous that I do not understand the ImageIO architecture well enough to support a custom ImageReader.

If ImageReader were an interface: I may feel comfortable using a Proxy interface implementation to intercept requests for a thumbnail. Or if com.sun.imageio.plugins.jpeg.JPEGImageReader was available to subclass I could try that. But the current architecture doesn't let me try either of those. For now I'll just do without.