Fixing QT Atom classes
I wrote a simple file that parsed all the Atoms in mp4s and the wrote the file back out. The output file sometimes wouldn't play back in QT Player.

I didn't isolate which changes here are strictly necessary. It may be the case that some changes here are necessary and some aren't. But with these changes the output file works.